infra/dev_dot_com/requirements-docker.txt
django==1.11.17
uwsgi==2.0.17.1
mysqlclient==1.3.14
django-reversion==3.0.3
django-reversion-compare==0.8.6
django-rest-hooks==1.5.0
django-rest-hooks-delivery
django-rest-swagger==2.2.0

celery>=4.0,<4.2  # ------------------------------------------------------------ (bumped 2018-06-07)
redis>=2.10,<2.11  # ----------------------------------------------------------- (bumped 2018-06-07)
celery-once>=1.2,<2.1  # ------------------------------------------------------- (bumped 2018-06-07)


# following is all stuff from silver's requirements/common.txt; minus the django requirement

# Please update the date when you decide a package's version
# NOT_LATEST means a newer is version available, but it was decided not to update
# UNMAINTAINED means the package owner is no longer maintaining it

# Core
# Django>=1.11,<2.0  # NOT_LATEST ----------------------------------------------- (checked 2018-06-07)
sqlparse>=0.2,<0.3  # (required for some old migrations) ----------------------- (bumped 2018-06-07)

# Django Utils
django-fsm>=2.3,<2.7  # -------------------------------------------------------- (bumped 2018-06-07)
django-filter>=1.0.4,<1.2  # --------------------------------------------------- (bumped 2018-06-07)
django-livefield>=2.8,<3.1  # -------------------------------------------------- (bumped 2018-06-07)
django-jsonfield==1.0.1  # ---------------------------------------------------- (checked 2018-06-07)
django-model-utils>=3.0,<3.2  # ------------------------------------------------ (bumped 2018-06-07)
django-annoying>=0.10,<0.11  # (various Django helpers) ------------------------ (bumped 2018-06-07)
django-autocomplete-light>=3.2,<3.3  # ----------------------------------------- (bumped 2018-06-07)

# API
djangorestframework>=3.6.3,<3.9  # --------------------------------------------- (bumped 2018-06-07)
djangorestframework-bulk<0.3  # ----------------------------------------------- (checked 2018-06-07)

# I18n
pycountry>=16.11.08  # --------------------------------------------------------- (bumped 2018-06-07)
python-dateutil>=2.6,<2.8  # --------------------------------------------------- (bumped 2018-06-07)
pyvat>=1.3,<1.4  # ------------------------------------------------------------- (bumped 2018-06-07)

# Crypto
cryptography>=1.9,<2.4  # ------------------------------------------------------ (bumped 2018-11-02)
PyJWT>=1.5,<1.7  # ------------------------------------------------------------- (bumped 2018-06-07)

# Other
furl>=1,<1.3  # (URL parsing and manipulation) --------------------------------- (bumped 2018-08-10)
xhtml2pdf>=0.2,<0.3  # (PDF rendering, python-dev is required) ----------------- (bumped 2018-06-07)
PyPDF2>=1.26,<2  # (PDF manipulation) ------------------------------------------ (bumped 2018-06-07)
